requirements.txt
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pypdf>=4.2.0
python-dotenv>=1.0.1
streamlit>=1.35.0
